feat: 부모 파트 연결된 가족 없음 뷰 구현, 이를 위한 홈 컴포넌트 구조 변경
